SUW: Remove classes and resources to create a minimal Setup Wizard
Brings the Setup Wizard down to the absolute minimum required in order to set up
a device and leave it in a provisioned state (which isn't always possible if the
Setup Wizard is simply omitted). Essentially, this leaves only the following:

- Bluetooth setup (which doesn't require user interaction)
- The final page that shows the logo and start button

All of the other settings from Setup Wizard (i.e. Wi-Fi, SIM, date/time, etc.)
can already be configured in Settings, so this allows the initial ROM setup to
be streamlined and let the user "get right into it."

Thanks to XDA user Aqq123, whose work on modifying the Setup Wizard script was
the inspiration for this patch (which goes much further by actually removing
all of Setup Wizard's unneeded classes and resources).